Phase 74: first-class corridor & threshold runtime
Replace the pair-shaped Place::Hallway {from,to} identity with stable
CorridorId identity, and resolve every crossing through
observed_facility::JunctionTopology::partner on an active socket set built
from the nav projection — never by reconstructing a room pair to decide
connectivity. Fuses the two vocabularies Phase 73 left parallel.

- teleport/mod.rs: drop the local ThresholdSlotId(u8), re-export the
  canonical observed_core ThresholdSlotId(u16); add Place::place_id/
  corridor_id and the deterministic corridor_id_for / corridor_socket_for
  derivations.
- teleport/transition.rs: place_junction builds the reciprocal room/corridor
  topology; apply_crossing resolves both directions via partner(); reshape
  Crossing::EnteredHallway to carry the CorridorId.
- teleport/geom.rs: enforce_active_sockets sources PlaceGeom.gaps from the
  active socket set, so render + Rapier apertures follow the topology (a
  sealed/unattached socket becomes a solid door). No-op on real data.
- teleport/test.rs: four regression fixtures for the "wall in front of a
  traversable threshold" class (traversable-never-walled, sealed-never-
  crossable, round-trip reciprocity, atomic reroute).
- sim/nav.rs, map_validation.rs: mechanical u8->u16 slot-cast updates.

Deferred to Phase 75 (documented shims kept compiling): the ~12 logic
consumers, the full corpus sweep, and unifying the runtime corridor-id space
with MapSpec::corridors edge indices. A pair-reconstruction fallback in
apply_crossing remains only for slotless authored/dev navs.

Verification: cargo fmt --check clean; clippy --workspace --all-targets 0
warnings; cargo test --workspace green; default-map determinism digest
unchanged.

+++ game/src/map_validation.rs
@@ -270,7 +270,7 @@ fn nav_for_spec_room(spec: &MapSpec, seed: u64, version: u32, room: RoomId) -> N
         .map(|(fallback, &target)| RoomConnectionSlot {
             target,
             slot: slot_for_connection(spec, room, target)
-                .unwrap_or(ThresholdSlotId(fallback as u8)),
+                .unwrap_or(ThresholdSlotId(fallback as u16)),
         })
         .collect();
     let corridor_roles = connections
@@ -328,7 +328,7 @@ fn edge_slot(edge: &MapEdge, room: RoomId) -> Option<ThresholdSlotId> {
 }
 
 fn slot_for_direction(direction: Direction) -> ThresholdSlotId {
-    ThresholdSlotId(direction.index() as u8)
+    ThresholdSlotId(direction.index() as u16)
 }
 
 fn audit_room_geom(
